backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0

# JWT认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis
redis==5.0.1
hiredis==2.2.3

# 异步数据库驱动
asyncpg==0.29.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
aiosqlite==0.19.0

# 性能测试
locust==2.18.3

# 代码质量
black==23.12.1
flake8==7.0.0
mypy==1.7.1
